+++ DeconTools.Backend/ProcessingTasks/Deconvoluters/HornDeconvolutor/ThrashV1/ChargeDetermination/AutoCorrelationChargeDetermination.cs
@@ -11,7 +11,7 @@ namespace DeconTools.Backend.ProcessingTasks.Deconvoluters.HornDeconvolutor.Thra
     public static class AutoCorrelationChargeDetermination
     {
         // might be too high a value.
-        private const short MaxCharge = 25;
+        private const int MaxCharge = 25;
 
         /// <summary>
         ///     Calculate the autocorrelation values for the an input.
@@ -55,7 +55,7 @@ namespace DeconTools.Backend.ProcessingTasks.Deconvoluters.HornDeconvolutor.Thra
         /// <param name="peak">is the peak whose charge we want to detect.</param>
         /// <param name="peakData">is the PeakData object containing raw data, peaks, etc which are used in the process.</param>
         /// <returns>Returns the charge of the feature.</returns>
-        public static short GetChargeState(ThrashV1Peak peak, PeakData peakData, bool debug)
+        public static int GetChargeState(ThrashV1Peak peak, PeakData peakData, bool debug)
         {
             var minus = 0.1;
             var plus = 1.1; // right direction to look
@@ -136,7 +136,7 @@ namespace DeconTools.Backend.ProcessingTasks.Deconvoluters.HornDeconvolutor.Thra
 
             // Determine the highest CS peak
             double bestAcScore;
-            short bestChargeState;
+            int bestChargeState;
             var success = HighestChargeStatePeak(minMz, maxMz, minN, autocorrelationScores, MaxCharge, out bestAcScore,
                 out bestChargeState);
 
@@ -148,7 +148,7 @@ namespace DeconTools.Backend.ProcessingTasks.Deconvoluters.HornDeconvolutor.Thra
             var charges = GenerateChargeStates(minMz, maxMz, minN, autocorrelationScores, MaxCharge, bestAcScore);
 
             // Get the final CS value to be returned
-            short returnChargeStateVal = -1;
+            int returnChargeStateVal = -1;
             // TODO: PattersonChargeStateCalculator really doesn't match the following code.
             var fwhm = peak.FWHM; // Store a copy of the FWHM to avoid modifying the actual value
             if (fwhm > 0.1)
@@ -220,10 +220,10 @@ namespace DeconTools.Backend.ProcessingTasks.Deconvoluters.HornDeconvolutor.Thra
         ///     Keep in mind that the autocorrelation score is not going to be perfectly at 0.5, or 0.33 etc, but approximately at
         ///     those points.
         /// </remarks>
-        private static List<short> GenerateChargeStates(double minMz, double maxMz, int minN, List<double> autocorrelationScores,
+        private static List<int> GenerateChargeStates(double minMz, double maxMz, int minN, List<double> autocorrelationScores,
             double maxChargeState, double bestAcScore)
         {
-            var chargeStates = new List<short>();
+            var chargeStates = new List<int>();
             // Preparation...
             var wasGoingUp = false;
 
@@ -246,7 +246,7 @@ namespace DeconTools.Backend.ProcessingTasks.Deconvoluters.HornDeconvolutor.Thra
                         // The default for Math.Round is MidpointRounding.ToEven, which will round 2.5 exactly to 2
                         // We should instead use MidpointRounding.AwayFromZero, which will round 2.5 to 3.
                         // TODO: This is a difference compared to PattersonChargeStateCalculator
-                        chargeStates.Add((short) (0.5 + chargeState));
+                        chargeStates.Add((int) (0.5 + chargeState));
                     }
                 }
                 wasGoingUp = goingUp;
@@ -266,7 +266,7 @@ namespace DeconTools.Backend.ProcessingTasks.Deconvoluters.HornDeconvolutor.Thra
         /// <param name="bestChargeState">returns the charge state at the best autocorrelation score distance, or -1 if none found.</param>
         /// <returns>true is successful, false otherwise</returns>
         private static bool HighestChargeStatePeak(double minMz, double maxMz, int minN, List<double> autocorrelationScores,
-            short maxChargeState, out double bestAutoCorrelationScore, out short bestChargeState)
+            int maxChargeState, out double bestAutoCorrelationScore, out int bestChargeState)
         {
             bestAutoCorrelationScore = -1;
             bestChargeState = -1;
@@ -283,7 +283,7 @@ namespace DeconTools.Backend.ProcessingTasks.Deconvoluters.HornDeconvolutor.Thra
                 var goingUp = autocorrelationScores[i] - autocorrelationScores[i - 1] > 0;
                 if (wasGoingUp && !goingUp)
                 {
-                    var chargeState = (short) (0.5 + numPts / ((maxMz - minMz) * (i - 1)));
+                    var chargeState = (int) (0.5 + numPts / ((maxMz - minMz) * (i - 1)));
                     var currentAutoCorScore = autocorrelationScores[i - 1];
                     if ((Math.Abs(currentAutoCorScore / autocorrelationScores[0]) > 0.05) && chargeState <= maxChargeState)
                     {
